Safari-611.1.10/LayoutTests/fast/text/zero-sized-fonts-expected.txt

PASS document.getElementById('sample').getBoundingClientRect().height < 30 is true
PASS successfullyParsed is true

TEST COMPLETE
This test makes sure that 0-sized fonts have 0-sized advances. The green rectangle below's height should be less than 30 pixels.
